Cast the purse for printing
player->au is int32_t and the format strings want an int.  The same width on
this platform, and not everywhere.

+++ src/cmd-wizard.c
@@ -1195,7 +1195,7 @@ void do_cmd_wiz_gain_gold(struct command *cmd)
 	else
 		player->au += (int32_t)amount;
 
-	msg("You have %d au.", player->au);
+	msg("You have %d au.", (int) player->au);
 
 	player->upkeep->redraw |= PR_GOLD;
 }
